Edit post fix and image preview

+++ blog/functions.py
@@ -7,6 +7,12 @@ from moviepy.editor import *
 
 src_szs = [480, 800, 1366, 1600, 1920]
 
+def validate_post_image(value):
+  ext = os.path.splitext(value.name)[1]
+  valid_extensions = ['.jpeg','.jpg','.bmp', '.png', '.tiff', '.gif', '.webm']
+  if not ext in valid_extensions:
+    raise ValidationError(u'File not supported!')
+
 
 def deleteThumb(text):
     # удаляем файлы картинок при удалении поста
@@ -55,31 +61,6 @@ def findFile(link):
         return ""
 
 
-"""def sidebarThumbnail(file_link, post_id):
-    ext = file_link.split('.')[-1].lower()
-    link = findLink(iri_to_uri(file_link))
-    file = uri_to_iri(findFile(link))
-    if ext == "gif":
-        file_out = uri_to_iri("/root/myblog/myblog/blog/static/media/{}/{}/{}/{}-{}-thumb.webm"\
-                    .format(link.group("year"), link.group("month"),
-                        link.group("day"), link.group("file"), str(post_id)))
-        file_out_gif = uri_to_iri("/tmp/{}-{}-thumb.gif"\
-                    .format(link.group("file"), str(post_id)))
-        link_out = uri_to_iri('/media/{}/{}/{}/{}-{}-thumb.webm'\
-                .format(link.group("year"), link.group("month"),
-                link.group("day"), link.group("file"), str(post_id)))
-        resize_gif(file, save_as=file_out_gif, resize_to=(150,150))
-        clip = VideoFileClip(file_out_gif)
-        video = CompositeVideoClip([clip])
-        video.write_videofile(file_out, codec='libvpx', fps=10,
-                                            audio=False, preset='superslow')
-        os.remove(file_out_gif)
-        return link_out
-    else:
-        return saveImage(link, file, 150, h=150)"""
-
-
-
 def saveImage(link, file, w, h=3500):
 
     file_out = '/root/myblog/myblog/blog/static/media/{}/{}/{}/{}-{}.{}'\
@@ -102,13 +83,6 @@ def saveImage(link, file, w, h=3500):
 def srcsets(text, wrap_a, post_id=None):
     """Make few srcsets"""
     soup = BeautifulSoup(uri_to_iri(text), "html5lib")  # текст поста
-    print("***************************")
-    print('soup ', str(soup))
-
-    # def not_processed_imgs(tag):
-    #    return (tag.name == 'img' and
-    #            'responsive-img' not in tag.get('class') and
-    #            'responsive-img,' not in tag.get('class'))
 
     img_links_raw = soup.find_all("img")  # ищем все картинки
     #img_links_raw = soup.find_all(not_processed_imgs)
@@ -149,13 +123,15 @@ def srcsets(text, wrap_a, post_id=None):
 
                 i['class'] = 'responsive-img post-image'
                 # если картинка больше нужного размера создаём миниатюру
-                w, h = Image.open(file).size
                 ext = i['src'].split('.')[-1].lower()
 
-                if ext == "gif":
+                if ext == "gif" or ext == "webm":
                     notgif = False
+                    clip = VideoFileClip(file)
+                    w, h = clip.size
                 else:
                     notgif = True
+                    w, h = Image.open(file).size
 
                 if notgif:
                     for sz in reversed(src_szs):
@@ -202,26 +178,29 @@ def srcsets(text, wrap_a, post_id=None):
                     i['sizes'] = "60vw"
 
                 else:  # конвертим гифки в webm
-                    file_out = uri_to_iri("/root/myblog/myblog/blog/static/media/{}/{}/{}/{}-{}.webm"\
-                    .format(link.group("year"), link.group("month"),
-                        link.group("day"), link.group("file"), str(post_id)))
-                    link_out = uri_to_iri('/media/{}/{}/{}/{}-{}.webm'\
-                            .format(link.group("year"), link.group("month"),
-                            link.group("day"), link.group("file"), str(post_id)))
-
-                    clip = VideoFileClip(file)
-                    video = CompositeVideoClip([clip])
-                    video.write_videofile(file_out, codec='libvpx',
-                                            audio=False, preset='superslow')
-                    os.remove(file)
-
                     webm = BeautifulSoup("", "html5lib").new_tag("video")
                     webm['autoplay'] = ""
                     webm['loop'] = ""
                     webm['controls'] = ""
                     webm['style'] = "max-width: " + str(w) + "px;"
                     source = BeautifulSoup("", "html5lib").new_tag("source")
-                    source['src'] = link_out
+                    if ext == "webm":
+                        source['src'] = "/media" + link.group()
+                    else:
+                        file_out = uri_to_iri("/root/myblog/myblog/blog/static/media/{}/{}/{}/{}-{}.webm"\
+                        .format(link.group("year"), link.group("month"),
+                            link.group("day"), link.group("file"), str(post_id)))
+                        link_out = uri_to_iri('/media/{}/{}/{}/{}-{}.webm'\
+                                .format(link.group("year"), link.group("month"),
+                                link.group("day"), link.group("file"), str(post_id)))
+
+                        clip = VideoFileClip(file)
+                        video = CompositeVideoClip([clip])
+                        video.write_videofile(file_out, codec='libvpx',
+                                                audio=False, preset='superslow')
+                        os.remove(file)
+
+                        source['src'] = link_out
                     source['type'] = "video/webm"
                     webm.insert(0, source)
                     i.replaceWith(webm)
@@ -235,113 +214,15 @@ def srcsets(text, wrap_a, post_id=None):
     return soup
 
 
-def resize_gif(path, save_as=None, resize_to=None):
-    """
-    Resizes the GIF to a given length:
-
-    Args:
-        path: the path to the GIF file
-        save_as (optional): Path of the resized gif. If not set, the original gif will be overwritten.
-        resize_to (optional): new size of the gif. Format: (int, int). If not set, the original GIF will be resized to
-                              half of its size.
-    """
-    all_frames = extract_and_resize_frames(path, resize_to)
-
-    if not save_as:
-        save_as = path
-
-    if len(all_frames) == 1:
-        print("Warning: only 1 frame found")
-        all_frames[0].save(save_as, optimize=True)
-    else:
-        all_frames[0].save(save_as, optimize=True, save_all=True, append_images=all_frames[1:], loop=1000)
-
-
-def analyseImage(path):
-    """
-    Pre-process pass over the image to determine the mode (full or additive).
-    Necessary as assessing single frames isn't reliable. Need to know the mode
-    before processing all frames.
-    """
-    im = Image.open(path)
-    results = {
-        'size': im.size,
-        'mode': 'full',
-    }
-    try:
-        while True:
-            if im.tile:
-                tile = im.tile[0]
-                update_region = tile[1]
-                update_region_dimensions = update_region[2:]
-                if update_region_dimensions != im.size:
-                    results['mode'] = 'partial'
-                    break
-            im.seek(im.tell() + 1)
-    except EOFError:
-        pass
-    return results
-
-
-def extract_and_resize_frames(path, resize_to=None):
-    """
-    Iterate the GIF, extracting each frame and resizing them
-
-    Returns:
-        An array of all frames
-    """
-    mode = analyseImage(path)['mode']
-
-    im = Image.open(path)
-
-    if not resize_to:
-        resize_to = (im.size[0] // 2, im.size[1] // 2)
-
-    i = 0
-    p = im.getpalette()
-    last_frame = im.convert('RGBA')
-
-    all_frames = []
-
-    try:
-        while True:
-            # print("saving %s (%s) frame %d, %s %s" % (path, mode, i, im.size, im.tile))
-
-            '''
-            If the GIF uses local colour tables, each frame will have its own palette.
-            If not, we need to apply the global palette to the new frame.
-            '''
-            if not im.getpalette():
-                im.putpalette(p)
-
-            new_frame = Image.new('RGBA', im.size)
-
-            '''
-            Is this file a "partial"-mode GIF where frames update a region of a different size to the entire image?
-            If so, we need to construct the new frame by pasting it on top of the preceding frames.
-            '''
-            if mode == 'partial':
-                new_frame.paste(last_frame)
-
-            new_frame.paste(im, (0, 0), im.convert('RGBA'))
-
-            new_frame.thumbnail(resize_to, Image.ANTIALIAS)
-            all_frames.append(new_frame)
-
-            i += 1
-            last_frame = new_frame
-            im.seek(im.tell() + 1)
-    except EOFError:
-        pass
-
-    return all_frames
-
 def stripMediaFromPath(file):
     """
     Strip "media/" from file's path
     """
-    path = file.split(os.sep)[2:]
-    new_path=""
-    for i in path:
-        new_path = os.path.join(new_path, i)
-    return uri_to_iri(new_path)
+    if re.search(r"^/media/", file):
+        path = file.split(os.sep)[2:]
+        new_path=""
+        for i in path:
+            new_path = os.path.join(new_path, i)
+        return uri_to_iri(new_path)
+    else:
+        return uri_to_iri(file)
